fixed bug in corr.axes so that it now uses the shared files bin numbers. finished adding subsample and tiers parameters to dist.shared.

+++ matrixoutputcommand.h
@@ -88,7 +88,7 @@ private:
 	};
 	vector<linePair> lines;
 	
-	void printSims(ostream&, vector< vector<float> >&);
+	void printSims(ostream&, vector< vector<double> >&);
 	int process(vector<SharedRAbundVector*>);
 	
 	vector<Calculator*> matrixCalculators;
